Add files via upload
2019-06-24 - I noticed the clearScreen and plotting functions left trash at 
the top of the display. In addition the first couple of pixels were not plotting
on the display (X and Y). Also the max pixels were not addressable leaving trash.
	        
To resole the issues I've created both X and Y offsets, tweaked the values so they
work (for me), set screen height to 128 and tweaked the clearScreen function.